Remove some unneeded rewriting code, including DComp. (#662)
* Remove the DComp custom comparator for the rewrite set (RSet).

3C may have relied on this custom order at one time, but it no longer
seems to; at least, the regression tests pass after this change.
Rewriting of non-overlapping declarations should be independent, and the
code already ensures that all rewrites to a given multi-decl are
performed together in the appropriate way.

RSet is now a std::map<Decl *, DeclReplacement *> so that
rewriteMultiDecl can look up the rewrite (if any) for a given Decl in
the multi-decl rather than scanning the set in the original custom
order. The use of a map adds some boilerplate, but I can't think of a
better option.

Also, VisitedMultiDeclMembers can just be a std::set<Decl *> because
that's all we need in order to test whether a Decl has already been
rewritten.

* The ParmVarDeclReplacement code is dead; remove it.

It seems that isPartOfFunctionPrototype is always true for a
ParmVarDecl and rewriting is handled as part of the containing
FunctionDecl instead. I haven't looked into when the old code path to
rewrite a single ParmVar became dead or whether it still works.

* hasFunctionBody in Utils.{h,cpp} is also dead; remove it.

I noticed this when searching for other code related to ParmVarDecl
that might be dead.