Fix for when a single file is the main file of multiple translation units (#666)
This is accomplished by identify the current translation unit using the index
of the translation unit in the vector of all translation units being
processed instead of name of the main file of the translation unit. This is
done for the primary expression map which previously used a (file name, node id)
pair as the key, and for the type variable map which previously used the
(file name, line, column) triple as the key. Both maps now use (TU index, node id).

In debugging the issue fixed by this the change, a few uninitialized fields were
discovered in PointerVaraibleConstraint. These did not cause the issue (or any
other concrete issue I have been able to find), but they have been initialized
anyways.

The constructor used for copying ConstraintVariables has been cleaned up and
refactored to use member initialization lists. The constructor is now used a few places
where a ConstraintVariable was previously constructed by manually extracting
fields from an existing ConstraintVariable into a new one.