Add other utility functions.
`IsSwitchCaseBlock`: use `dyn_cast_or_null` to cover the null pointer
case.

`ConditionOnEdge`: do not test if there is no edge between
pred to curr since it will only be called if there is an edge.

`GetModifiedVars`: use `TranspareCasts` to bypass some casting.
The feature to deal with membership access and the array indexing is
still TODO.