Change test7 in functionDeclEnd.c to infer a valid bound
The purpose of this function was to test function declaration rewriting
for itype array pointers declared with bounds but without an explicit
itype. This is equally well tested when a correct bound is inferred for
the parameter array.

(Matt: Remove the XFAIL now that the test is fixed.)

(cherry picked from commit 57d31d21ccf25fb1bd45e06033036431c0a0cc47)