Don't add `void` prototypes on functions that would otherwise not need to be rewritten (#688)
A function that does not otherwise need to be rewritten would sometimes be
rewritten in order to replace a missing prototype with (void). This can trigger
a rewriting error if the function is defined by a macro or in an unwritable
file.

After these changes, 3C still adds a prototype for functions with checked or
itype return types, so int `*foo() { return 0; }` will still convert to
`_Ptr<int> foo(void) { return 0; }` as is required by Checked C.

Checked C also requires that a functions has a prototype for it to be called
from inside a checked scope, but this is not handled by 3C; see issue
correctcomputation/checkedc-clang#382.  The status of this issue has not
changed.