Ensure bounds are not inserted into unwritable files (#689)
Adds checks to ensure that 3C will not attempt to rewrite in unwritable files
to insert array bounds. The test file in this PR contains an example using a
function body in an unwritable file which fails on the current main. After
changes made to support rewriting in functions declarations without
definitions (correctcomputation/checkedc-clang#691), this will be able to
happen even when the functions are not defined, and so it will becomes a more
prevalent issue.