Insert itypes for parameters with typedef types (#690)
This allows a function using an otherwise unchecked typedef type to expose an
checked interface. It also lets a function use a typedef'ed parameter unsafely
without forcing the entire typedef to be unchecked.

In the following example, the typedef td is unchecked because of the unsafe use
of b. Prior to this change, test would not be rewritten, leaving it with a
fully unchecked type. The new rewriting is `void test(td a : itype(_Ptr<int>))`.

    typedef int *td;
    void test(td a) {
      td b = 1;
    }

In the other direction, the parameter can be used unsafely, but the typedef
will still be rewritten to `_Ptr<int>`, and the parameter will get an itype
`int * : itype(td)`.

    typedef int *td;
    void test(td a) {
      a = 1;
    }