Generate itype string from AST when extracting from source fails (#693)
Fixes an assertion failure (correctcomputation/checkedc-clang#594) when
attempting to extract the string representation of an itype expression from the
original source code if the itype expression is inside a macro. The string
representation for itypes in macros is now re-generated from the AST.

The fix in its current form leads to some unnecessary expansion of macros. This
is discussed in correctcomputation/checkedc-clang#694.