Insert itypes correctly for constant sized arrays (#702)
Given the declaration of a global variable `int foo[10]` converted with
`3c -itypes-for-extern -alltypes` the resulting declaration incorrectly moved
the length of the array before identifier. This changes fixes the rewriting to
correctly be `int foo[10] : itype(int _Checked[10])`. This also fixes itype
insertion for pointers to constant size arrays.

Another change is included to pass idempotence checks for the new tests added.
This change expands the source range replaced for global variable declarations
to include the itype or bounds declarations if present.