Treat the bounds of _Nt_checked as one less than the written length (#653)
The Checked C bounds for constant sized `_Nt_checked` arrays is one less
than the written size because the written size counts the null
terminator while the Checked C bounds do not.

To make this change, the declared bounds of constant size arrays must be
added to the array bounds inference information after pointer type
constraint solving. This lets `_Checked` and `_Nt_checked` array be treated
differently.

This also fixes an error in `canBeNtArray` that caused  `ensureNtCorrect `to add
incorrect constraints. The function checked if a type was a `PointerType` or
`ArrayType`, but it could be a wrapper type around one of these types. For
example, `ParenType` and `DecayedType` instances were both treated as
not pointers or arrays even if the type they wrapped was a pointer or array type.

Fixes correctcomputation/checkedc-clang#396