Fix array bounds heuristics on itype parameters (#654)
The internal constraint variable for function parameters was used to checked the
solved pointer type when adding array bounds heuristics.  This made itype
parameters look like wild pointers instead of array or nt array pointers, so
the heuristics were not applied. Heuristics are now applied based on the
external parameter, which solves to a checked type for itype parameters.

This resolves the remaining issue in correctcomputation/checkedc-clang#487