Fix 3c tests failing when check-all is run.
The 3C developers added a separate test target for 3c called check-3c. This
sped up testing 3c. However, the clang build also adds a test target for each
clang-based tool. This results in two targets in check-all that both run 3c tests.
This in turn leads to 2 builds of the 3C tests suite being run in parallel when
check-all is run in parallel. This causes things such as file-level access conflicts,
which causes some 3C tests to fail.

This changes removes the separate test target for 3c, so that 3c tests follow the
same approaches used for tests for other clang-based tools in the repo.