Implement alternate syntax and macros for backward compatibility..
This change implements alternate syntax and macros to enhance the backward compatibility of Checked C code with non-Checked C compilers. The technical details of the changes are based on the proposal documented in [this wiki page](https://github.com/secure-sw-dev/checkedc/wiki/Proposed-extension-changes-to-improve-backward-compatibility). The key changes include:

- Pointer types: Introduce _Array, _Nt_array, and _Single qualifiers for pointer types, which can be defined as empty strings when Checked C is not supported.
- Arrays: Define the keyword _Checked/_Nt_checked as an empty string when a compiler does not support Checked C.
- Generic types: Introduce the _TyArg macro to wrap type arguments, and the _TyVar macro for type variables. Also, provide a two-part approach for generic function definitions and generic struct definitions.
- Bounds annotations: Introduce new keywords _Any, _Bounds, _Byte_count, and _Count for bounds expressions, and allow a declarator to be followed by a bounds expression.
- Dynamic check expressions: Implement a macro that evaluates the argument but does nothing with it, or alternatively, make use of the assert macro.
- Bounds cast expressions: Introduce variadic macros for _Dynamic_bounds_cast and _Assume_bounds_cast expressions, which map to their corresponding syntax forms when Checked C is supported or to a C-style cast when it is not.
- Where clauses: Allow optional parentheses to surround the where clause, and implement a macro that discards the body of the where clause, when Checked C, is not supported.

These changes enable Checked C code to be compiled with non-Checked C compilers while preserving the benefits and safety guarantees when using a Checked C compiler. This improves code portability and broadens the range of platforms on which Checked C code can run.

Testing:
Changed the Checked C tests to use the new syntax in some existing tests.

Passes all Checked C cand clang tests on 64-bit Windows 11 and Ubuntu 20.03.  For all tests, fails 1 test on each platform.  The failures appear completely unrelated to this change.