Fix crash when processing GCC statement expression (#1226)
This fixes issue #1220, a crash when processing a GCC statement expression. The problem was that lexicographic comparison of expressions assumed that all children of expressions are expressions also. This is no true for the GCC statement expression extension, which wraps a statement in an expression.    This change includes a new regression test case for the issue.