Fix merge errors causing test failures or crashes.
This changes fixes three issues:
1. The mainline clang sources changed the type of BuildDeclRef
to  DeclRef * from of an ExprResult. When we did the merge of
upstream, we handled this incorrectly. This caused the
compiler to crash on  bounds-safe interface.

2. Some tests fail because the error messages around checked arrays
are missing a space before the "_Checked" keyword.

3. Update testing instructions.

For BuildDeclRef, in the merge tried casting the expression to
a DeclRefExpr before returning it.  This doesn't work because
we use BuildDeclRef in Checked C to introduce casts for
bounds-safe interfaces. The casts are introduced when
variables with bounds-safe interfacs are accessed in checked scopes.

The return type used to be an ExprResult, which handled
a implicit cast expr being returned fine for the most
part. ExprResult on Expr * that indicates either a valid expression
was created or  the expression is invalid.  It allows any Expr * to be
returned but the caller may only be expecting a specific Expr subclass.

To fix, this change the return type of BuildDeclRefExpr to be an
Expr *. For callers that  must have a DeclRefExpr (which is rare), add
code that dynamically downcasts the type. The downcasts are only added
for OpenMP and C++ overload resolution. Checked C features are not
supported for C++, so this is a non-issue. For Open MP, it is
possible that if put a bounds-safe interface on a loop variable
that is important to OpenMP, the compiler might crash.  The
fix is to NOT do that.  I added an assert for now.  We could
turn this into an error.