Fix some Checked C clang errors.
- mergeTypes now takes an additional paramter related to conditional
opeators.  The ignoreBounds flag is last on the command parameter list
so we need to add some additional parameters in cases where it has been
set.
- There are Checked C errors that can trigger a general diagonstic
complaint from clang, such as "mismatched types".  To avoid
redundant error messages, we try to diagonse Checked C errors first
and exit before reaching general diagnostic code.  There's a new
case where we need to do that related to clang having stricer.
checking of no-prototype functions.

The result of running check-clang is that we have on x64 windows is:
- 84 tests pass
- 9 tests fail.
- 6 tests are unsupported.