Update how driver sets the Checked C language option.
The way we set the Checked C language option in the driver
is not working any more; the flag is being ignored as
language flag handling has moved to a more declarative
approach.

There are still two flags:
- -fcheckedc-extension
- -fno-checkedc-extension.

Enabled the Checked C extension by default for the
compiler when the language choice is C.  Don't pass
the [-fcheckedc-extension flag from the driver to
the core compiler.  Do pass the -fno-checkedc-extension
flag to the core compiler invocation.

This approach works, but the one downsie is that we
currently don't warn if you try to combined the
-fcheckedc-extension with other languages.  The driver
just silently ignores the flag.