Fix Checked C driver flags
My previous attempt at fixing this did not work.  The driver
was always setting the Checked C language option based on
command-line flags.  Since the driver doesn't really understand
what language is being compiled, this does not work.

We want the following:
- For C, the Checked C extension is enabled by default.
- For other language it is off.
- There are two flags -fcheckedc-extension and
  -fno-checkedc-extension.
- If the first flag (-fcheckedc-extension) is passed to cc1,
  nothing is done for C, since it is the default. A warning
  is issued for other languages and it ignored.
- The second flag disabled the Checked C extension.

One tricky thing that I found is that cc1 generates a new
command-line flag and uses that, in an attempt to detect
driver non-determinism.  So when you add a new flag, you
have to go to CompilerInvocation::GenerateLangArgs and make sure
your new setting results in a flag being generated there.
Otherwise your new flag is ignored.  This is not documented
well and took me a few hours to figure out.