Update compiler flags to use marshalling infrastructure.
There are compiler flags for dumping Checked C-related
internal data structures.  Update them to be handle
via the clang/cc1 marshalling infrastructure, instead
of handling them manually. This addresses the issue
that the clang driver silently is throwing away the
flags, which in turns causes a bunch of tests that
are setting the flags to fail.